requirements.txt
boto3==1.26.71
botocore==1.29.71
gitdb==4.0.10
GitPython==3.1.30
jmespath==1.0.1
python-dateutil==2.8.2
python-git==2018.2.1
python-terraform==0.10.1
s3transfer==0.6.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
urllib3==1.26.14
